IBM Introduction to ML Specialization/Deep Learning & Reinforcement Learning/Week6/Brain_Tumor_Detector_Deep_Learning-master/requirements.txt
Flask==2.0.1
Jinja2==3.0.3
numpy==1.19.5
opencv-python==4.1.1.26
pandas==1.1.3
scikit-learn==0.24.1
tensorflow-gpu==2.3.0rc0
Werkzeug==2.0.1
